refactor: move duplicate code into super class
